Bugfixes and improved DASH support
1. Fixed issue with possible deadlock condition when sync thread needs to restart due to source stream discontinuity
2. Improved support for DASH segmenttemplate/segmenttimeline- works well with one stream on some players but not all so I am still trying to do the interop with the dashif reference player.
It is likely the interop/playback issue is related to timing or mpd file parameters.
Using symlinks against the generated fragmented mp4 files for the segmenttimeline.  The fragmented mp4s are generated with a sequence number (for the fragmented fmp4 based hls),
but this is not used for dash.  I was trying to use one set of files for both.
My next step is to figure out Youtube DASH based publishing as well as bitstream switching using a single init.mp4 for audio/video.  There is a slightly different syntax for this.

Trying to debug playback of dash with some additional players.  Reviewing specs to figure out where the discrepency might be.